requirements.txt
absl-py==0.9.0
appnope==0.1.0
astor==0.8.1
astroid==2.3.3
autopep8==1.5
backcall==0.1.0
Click==7.0
cycler==0.10.0
Cython==0.29.15
decorator==4.4.1
Flask==1.1.1
gast==0.2.2
google-pasta==0.1.8
grpcio==1.27.1
h5py==2.10.0
ipython==5.1.0
ipython-genutils==0.2.0
isort==4.3.21
itsdangerous==1.1.0
jedi==0.16.0
Jinja2==2.11.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
lazy-object-proxy==1.4.3
Markdown==3.2.1
MarkupSafe==1.1.1
matplotlib==1.5.3
mccabe==0.6.1
numpy==1.16.2
opencv-python==3.4.2.16
opt-einsum==3.1.0
parso==0.6.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.0.0
prompt-toolkit==1.0.18
protobuf==3.11.3
ptyprocess==0.6.0
pycodestyle==2.5.0
Pygments==2.5.2
pylint==2.4.4
pyparsing==2.4.6
pytesseract==0.3.2
python-dateutil==2.8.1
pytz==2019.3
rope==0.16.0
Shapely==1.7.0
simplegeneric==0.8.1
six==1.14.0
tensorboard==1.15.0
tensorflow==1.15.0
tensorflow-estimator==1.15.1
termcolor==1.1.0
tqdm==4.11.2
traitlets==4.3.3
typed-ast==1.4.1
UNKNOWN==0.0.0
wcwidth==0.1.8
Werkzeug==1.0.0
wrapt==1.11.2
